Update Colorduino.cpp - Increased Brightness
Moved the 'close_all_lines' command from the ISR to inside the 'Colorduino.run' function, near the end, just after transferring all data to shift registers, and just before setting the latch 'LED_LAT_SET' or in the DM163 spec 'LAT_B'.

Previously, we were turning the LEDs off, then transferring all the data to the DM163 shift registers, then latching that data, then enabling the specific line again. This meant that the LEDs were 'off' for a significant amount of time while we were loading the shift registers, resulting in a display that was significantly dimmer than it needed to be.

As far as I can tell from the DM163 spec, the output will not change until the data is latched, so it is safe to leave the previous line of LEDs active right up until we latch the data for the next row.

Moved resetting 'TCNT2' and 'TCNT4' to be the last command in the ISR, otherwise it would not be an accurate timer, as we have seen that our ISR function itself does take time to complete. From this point of view, the value 0x61 may need to be changed, as